重写thread_page，把post_input打包成单文件组件
