+++ resources/vue/component/emoji.vue
