+++ resources/vue/component/post_item.vue
